Split out Mixpanel and MixpanelPeople, add class extensions to Private.h header files.
