+++ Mixpanel/MixpanelPeoplePrivate.h
@@ -1,0 +1,22 @@
+//
+//  MixpanelPeoplePrivate.h
+//  Mixpanel
+//
+//  Created by Sam Green on 6/16/16.
+//  Copyright © 2016 Mixpanel. All rights reserved.
+//
+#import "MixpanelPeople.h"
+
+@class Mixpanel;
+
+@interface MixpanelPeople ()
+
+@property (nonatomic, weak) Mixpanel *mixpanel;
+@property (nonatomic, strong) NSMutableArray *unidentifiedQueue;
+@property (nonatomic, copy) NSString *distinctId;
+@property (nonatomic, strong) NSDictionary *automaticPeopleProperties;
+
+- (instancetype)initWithMixpanel:(Mixpanel *)mixpanel;
+- (void)merge:(NSDictionary *)properties;
+
+@end
